+++ NEWS.md
@@ -1,5 +1,7 @@
 # CopernicusClimate v0.0.2.0004
 
+* New feature:
+  * Translator for Python request: `cds_python_to_r()`
 * Updated documentation
 * Fixes in `cds_build_request()`
 * improved test coverage